CF-138 Refactor generate_load.py
Moved BasePrerequisites to base.py
Moved CleanEnv to cleanup.py
Reworked Prerequisites in generate_load.py
Moved all tests to tests/ dir
Added lost licenses.
