+++ devlab/tests/generate_load.py
@@ -1,24 +1,32 @@
+# Copyright 2015 Mirantis Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import argparse
 import itertools
 import time
 import json
 import os
 import yaml
-import config as conf
-import utils
 import ConfigParser
 
-from glanceclient.v1 import Client as glance
 from keystoneclient import exceptions as ks_exceptions
-from keystoneclient.v2_0 import client as keystone
 from neutronclient.common import exceptions as nt_exceptions
-
 from novaclient import exceptions as nv_exceptions
-from novaclient.v1_1 import client as nova
-from neutronclient.v2_0 import client as neutron
-from cinderclient.v1 import client as cinder
 
-from test_exceptions import NotFound
+from base import BasePrerequisites
+from cleanup import CleanEnv
+import config as conf
 
 TIMEOUT = 600
 VM_SPAWNING_LIMIT = 5
@@ -29,9 +37,6 @@ CREATE_CLEAN_METHODS_MAP = {
     'create_flavors': 'clean_flavors',
     'create_all_networking': 'clean_all_networking'
 }
-OPENSTACK_RELEASES = {'192.168.1.2': 'grizzly',
-                      '192.168.1.3': 'icehouse',
-                      '192.168.1.8': 'juno'}
 
 
 def clean_if_exists(func):
@@ -76,259 +81,6 @@ def get_dict_from_config_file(config_file):
         for option in config_file.options(section):
             conf_dict[section][option] = config_file.get(section, option)
     return conf_dict
-
-
-class BasePrerequisites(object):
-
-    def __init__(self, config, configuration_ini, cloud_prefix='SRC'):
-        self.configuration_ini = configuration_ini
-        self.filtering_utils = utils.FilteringUtils(
-            self.configuration_ini['migrate']['filter_path'])
-        self.migration_utils = utils.MigrationUtils(config)
-
-        self.config = config
-        self.cloud_prefix = cloud_prefix.lower()
-
-        self.username = self.configuration_ini[self.cloud_prefix]['user']
-        self.password = self.configuration_ini[self.cloud_prefix]['password']
-        self.tenant = self.configuration_ini[self.cloud_prefix]['tenant']
-        self.auth_url = self.configuration_ini[self.cloud_prefix]['auth_url']
-
-        self.keystoneclient = keystone.Client(auth_url=self.auth_url,
-                                              username=self.username,
-                                              password=self.password,
-                                              tenant_name=self.tenant)
-        self.keystoneclient.authenticate()
-
-        self.novaclient = nova.Client(username=self.username,
-                                      api_key=self.password,
-                                      project_id=self.tenant,
-                                      auth_url=self.auth_url)
-
-        self.token = self.keystoneclient.auth_token
-
-        self.image_endpoint = \
-            self.keystoneclient.service_catalog.get_endpoints(
-                service_type='image',
-                endpoint_type='publicURL')['image'][0]['publicURL']
-
-        self.glanceclient = glance(endpoint=self.image_endpoint,
-                                   token=self.token)
-
-        self.neutronclient = neutron.Client(username=self.username,
-                                            password=self.password,
-                                            tenant_name=self.tenant,
-                                            auth_url=self.auth_url)
-
-        self.cinderclient = cinder.Client(self.username, self.password,
-                                          self.tenant, self.auth_url)
-        self.openstack_release = self._get_openstack_release()
-        self.server_groups_supported = self.openstack_release in ['icehouse',
-                                                                  'juno']
-
-    def _get_openstack_release(self):
-        for release in OPENSTACK_RELEASES:
-            if release in self.auth_url:
-                return OPENSTACK_RELEASES[release]
-        raise RuntimeError('Unknown OpenStack release')
-
-    def get_vagrant_vm_ip(self):
-        for release in OPENSTACK_RELEASES:
-            if release in self.auth_url:
-                return release
-
-    def get_tenant_id(self, tenant_name):
-        for tenant in self.keystoneclient.tenants.list():
-            if tenant.name == tenant_name:
-                return tenant.id
-        raise NotFound('Tenant with name "%s" was not found' % tenant_name)
-
-    def get_tenant_name(self, tenant_id):
-        for tenant in self.keystoneclient.tenants.list():
-            if tenant.id == tenant_id:
-                return tenant.name
-        raise NotFound('Tenant with id "%s" was not found' % tenant_id)
-
-    def get_user_id(self, user_name):
-        for user in self.keystoneclient.users.list():
-            if user.name == user_name:
-                return user.id
-        raise NotFound('User with name "%s" was not found' % user_name)
-
-    def get_router_id(self, router):
-        _router = self.neutronclient.list_routers(name=router)['routers']
-        if _router:
-            return _router[0]['id']
-        raise NotFound('Router with name "%s" was not found' % router)
-
-    def get_image_id(self, image_name):
-        for image in self.glanceclient.images.list(is_public=None):
-            if image.name == image_name:
-                return image.id
-        raise NotFound('Image with name "%s" was not found' % image_name)
-
-    def get_flavor_id(self, flavor_name):
-        for flavor in self.novaclient.flavors.list():
-            if flavor.name == flavor_name:
-                return flavor.id
-        raise NotFound('Flavor with name "%s" was not found' % flavor_name)
-
-    def get_vm_id(self, vm_name):
-        for vm in self.novaclient.servers.list(search_opts={'all_tenants': 1}):
-            if vm.name == vm_name:
-                return vm.id
-        raise NotFound('VM with name "%s" was not found' % vm_name)
-
-    def get_role_id(self, role_name):
-        for role in self.keystoneclient.roles.list():
-            if role.name == role_name:
-                return role.id
-        raise NotFound('Role with name "%s" was not found' % role_name)
-
-    def get_net_id(self, net):
-        _net = self.neutronclient.list_networks(
-            name=net, all_tenants=True)['networks']
-        if _net:
-            return _net[0]['id']
-        raise NotFound('Network with name "%s" was not found' % net)
-
-    def get_net_name(self, net_id):
-        _net = self.neutronclient.list_networks(id=net_id,
-                                                all_tenants=True)['networks']
-        if _net:
-            return _net[0]['name']
-        raise NotFound('Network with id "%s" was not found' % net_id)
-
-    def get_sg_id(self, sg):
-        _sg = self.neutronclient.list_security_groups(
-            name=sg, all_tenants=True)['security_groups']
-        if _sg:
-            return _sg[0]['id']
-        raise NotFound('Security group with name "%s" was not found' % sg)
-
-    def get_server_group_id(self, server_group_name):
-        for server_group in self.get_all_server_groups():
-            if server_group.name == server_group_name:
-                return server_group.id
-        msg = 'Server group with name "%s" was not found'
-        raise NotFound(msg % server_group_name)
-
-    def get_volume_id(self, volume_name):
-        volumes = self.cinderclient.volumes.list(
-            search_opts={'all_tenants': 1})
-        for volume in volumes:
-            if volume.display_name == volume_name:
-                return volume.id
-        raise NotFound('Volume with name "%s" was not found' % volume_name)
-
-    def get_volume_snapshot_id(self, snapshot_name):
-        snapshots = self.cinderclient.volume_snapshots.list(
-            search_opts={'all_tenants': 1})
-        for snapshot in snapshots:
-            if snapshot.display_name == snapshot_name:
-                return snapshot.id
-        raise NotFound('Snapshot with name "%s" was not found' % snapshot_name)
-
-    def get_user_tenant_roles(self, user):
-        user_tenant_roles = []
-        for tenant in self.keystoneclient.tenants.list():
-            user_tenant_roles.extend(self.keystoneclient.roles.roles_for_user(
-                user=self.get_user_id(user.name),
-                tenant=self.get_tenant_id(tenant.name)))
-        return user_tenant_roles
-
-    def get_ext_routers(self):
-        routers = self.neutronclient.list_routers()['routers']
-        ext_routers = [router for router in routers
-                       if router['external_gateway_info']]
-        return ext_routers
-
-    def get_sec_group_id_by_tenant_id(self, tenant_id):
-        sec_group_list = self.neutronclient.list_security_groups()
-        return [i['id'] for i in sec_group_list['security_groups']
-                if i['tenant_id'] == tenant_id]
-
-    def get_users_keypairs(self):
-        self.switch_user(self.username, self.password, self.tenant)
-        user_names = [u['name'] for u in self.config.users]
-        keypairs = []
-        for user in self.config.users:
-            if user['name'] not in user_names:
-                continue
-            if not user.get('enabled') or user.get('deleted'):
-                continue
-            if not self.tenant_exists(user['tenant']) or \
-                    not self.user_has_not_primary_tenants(user['name']):
-                continue
-            try:
-                self.switch_user(user['name'], user['password'],
-                                 user['tenant'])
-            except ks_exceptions.Unauthorized:
-                self.keystoneclient.users.update(
-                    self.get_user_id(user['name']), password=user['password'],
-                    tenant=user['tenant'])
-                self.switch_user(user['name'], user['password'],
-                                 user['tenant'])
-            keypairs.extend(self.novaclient.keypairs.list())
-        return keypairs
-
-    def get_all_server_groups(self):
-        initial_tenant = self.keystoneclient.tenant_name
-        self.switch_user(self.username, self.password, self.tenant)
-        server_groups = self.novaclient.server_groups.list()
-        for tenant in self.config.tenants:
-            if not self.tenant_exists(tenant['name']):
-                continue
-            with utils.AddAdminUserRoleToNonAdminTenant(
-                    self.keystoneclient, self.username, tenant['name']):
-                self.switch_user(self.username, self.password,
-                                 tenant['name'])
-                server_groups.extend(self.novaclient.server_groups.list())
-        self.switch_user(self.username, self.password, initial_tenant)
-        return server_groups
-
-    def user_has_not_primary_tenants(self, user_name):
-        user_id = self.get_user_id(user_name)
-        for tenant in self.keystoneclient.tenants.list():
-            if self.keystoneclient.roles.roles_for_user(user=user_id,
-                                                        tenant=tenant.id):
-                return True
-        return False
-
-    def check_vm_state(self, srv):
-        srv = self.novaclient.servers.get(srv)
-        return srv.status == 'ACTIVE'
-
-    def tenant_exists(self, tenant_name):
-        self.switch_user(self.username, self.password, self.tenant)
-        try:
-            self.get_tenant_id(tenant_name)
-            return True
-        except NotFound:
-            return False
-
-    def switch_user(self, user, password, tenant):
-        self.keystoneclient = keystone.Client(auth_url=self.auth_url,
-                                              username=user,
-                                              password=password,
-                                              tenant_name=tenant)
-        self.keystoneclient.authenticate()
-        self.token = self.keystoneclient.auth_token
-        self.novaclient = nova.Client(username=user,
-                                      api_key=password,
-                                      project_id=tenant,
-                                      auth_url=self.auth_url)
-
-        self.glanceclient = glance(endpoint=self.image_endpoint,
-                                   token=self.token)
-
-        self.neutronclient = neutron.Client(username=user,
-                                            password=password,
-                                            tenant_name=tenant,
-                                            auth_url=self.auth_url)
-
-        self.cinderclient = cinder.Client(user, password, tenant,
-                                          self.auth_url)
 
 
 class Prerequisites(BasePrerequisites):
@@ -1260,323 +1012,6 @@ class Prerequisites(BasePrerequisites):
         self.create_user_on_dst()
         print('>>> Creating networks on dst:')
         self.create_network_with_segm_id()
-
-
-class CleanEnv(BasePrerequisites):
-
-    def clean_vms(self):
-        vms = self.migration_utils.get_all_vms_from_config()
-        vms_names = [vm['name'] for vm in vms]
-        vms = self.novaclient.servers.list(search_opts={'all_tenants': 1})
-        vms_ids = []
-        for vm in vms:
-            if vm.name not in vms_names:
-                continue
-            vms_ids.append(vm.id)
-            self.novaclient.servers.delete(vm.id)
-            print('VM "%s" has been deleted' % vm.name)
-        self.wait_vms_deleted()
-
-    def wait_vms_deleted(self, tenant_id=None):
-        search_opts = {'all_tenants': 1}
-        if tenant_id is not None:
-            search_opts['tenant_id'] = tenant_id
-        timeout = 120
-        for _ in range(timeout):
-            servers = self.novaclient.servers.list(
-                search_opts=search_opts)
-            if not servers:
-                break
-            for server in servers:
-                if server.status != 'DELETED':
-                    time.sleep(1)
-                try:
-                    self.novaclient.servers.delete(server.id)
-                except nv_exceptions.NotFound:
-                    pass
-        else:
-            raise RuntimeError('Next vms were not deleted')
-
-    def clean_volumes(self):
-        volumes = self.config.cinder_volumes
-        volumes += itertools.chain(*[tenant['cinder_volumes'] for tenant
-                                     in self.config.tenants if 'cinder_volumes'
-                                     in tenant])
-        for volume in self.config.cinder_volumes_from_images:
-            volumes.append(volume)
-        volumes_names = [volume['display_name'] for volume in volumes]
-        volumes = self.cinderclient.volumes.list(
-            search_opts={'all_tenants': 1})
-        for volume in volumes:
-            if volume.display_name not in volumes_names:
-                continue
-            self.cinderclient.volumes.delete(
-                self.get_volume_id(volume.display_name))
-            print('Volume "%s" has been deleted' % volume.display_name)
-
-    def clean_flavors(self):
-        flavors_names = [flavor['name'] for flavor in self.config.flavors]
-        for tenant in self.config.tenants:
-            if not tenant.get('flavors'):
-                continue
-            flavors_names.extend([f['name'] for f in tenant['flavors']])
-        for flavor in self.novaclient.flavors.list(is_public=None):
-            if flavor.name not in flavors_names:
-                continue
-            self.novaclient.flavors.delete(self.get_flavor_id(flavor.name))
-            print('Flavor "%s" has been deleted' % flavor.name)
-
-    def clean_images(self):
-        all_images = self.migration_utils.get_all_images_from_config()
-        images_names = [image['name'] for image in all_images]
-
-        for image in self.glanceclient.images.list(is_public=None):
-            if image.name not in images_names:
-                continue
-            self.glanceclient.images.delete(self.get_image_id(image.name))
-            print('Image "%s" has been deleted' % image.name)
-
-    def clean_snapshots(self):
-        snaps_names = [snapshot['image_name']
-                       for snapshot in self.config.snapshots]
-        for snapshot in self.glanceclient.images.list():
-            if snapshot.name not in snaps_names:
-                continue
-            self.glanceclient.images.delete(
-                self.get_image_id(snapshot.name))
-            print('Snapshot "%s" has been deleted' % snapshot.name)
-
-    def clean_networks(self):
-        nets = self.config.networks
-        nets += itertools.chain(*[tenant['networks'] for tenant
-                                  in self.config.tenants
-                                  if tenant.get('networks')])
-        nets_names = [net['name'] for net in nets]
-        for network in self.neutronclient.list_networks()['networks']:
-            if network['name'] not in nets_names:
-                continue
-            net_id = self.get_net_id(network['name'])
-            self.clean_network_ports(net_id)
-            self.neutronclient.delete_network(net_id)
-            print('Network "%s" has been deleted' % network['name'])
-
-    def delete_port(self, port):
-        port_owner = port['device_owner']
-        if port_owner == 'network:router_gateway':
-            self.neutronclient.remove_gateway_router(port['device_id'])
-        elif port_owner == 'network:router_interface':
-            self.neutronclient.remove_interface_router(
-                port['device_id'], {'port_id': port['id']})
-        elif port_owner == 'network:dhcp' or not port_owner:
-            self.neutronclient.delete_port(port['id'])
-        elif 'LOADBALANCER' in port_owner:
-            vips = self.neutronclient.list_vips(port_id=port['id'])['vips']
-            for vip in vips:
-                self.neutronclient.delete_vip(vip['id'])
-        else:
-            msg = 'Unknown port owner %s'
-            raise RuntimeError(msg % port['device_owner'])
-
-    def clean_network_ports(self, net_id):
-        ports = self.neutronclient.list_ports(network_id=net_id)['ports']
-        for port in ports:
-            self.delete_port(port)
-
-    def clean_router_ports(self, router_id):
-        ports = self.neutronclient.list_ports(device_id=router_id)['ports']
-        for port in ports:
-            self.delete_port(port)
-
-    def clean_routers(self):
-        router_names = [router['router']['name']
-                        for router in self.config.routers]
-        for tenant in self.config.tenants:
-            if tenant.get('routers'):
-                for router in tenant['routers']:
-                    router_names.append(router['router']['name'])
-        for router in self.neutronclient.list_routers()['routers']:
-            if router['name'] not in router_names:
-                continue
-            router_id = self.get_router_id(router['name'])
-            self.clean_router_ports(router_id)
-            self.neutronclient.delete_router(router_id)
-            print('Router "%s" has been deleted' % router['name'])
-
-    def clean_fips(self):
-        floatingips = self.neutronclient.list_floatingips()['floatingips']
-        for ip in floatingips:
-            try:
-                self.neutronclient.delete_floatingip(ip['id'])
-            except nt_exceptions.NeutronClientException as e:
-                print "Ip %s failed to delete: %s" % (
-                    ip['floating_ip_address'], repr(e))
-
-    def clean_security_groups(self):
-        sgs = self.neutronclient.list_security_groups()['security_groups']
-        for sg in sgs:
-            try:
-                self.neutronclient.delete_security_group(self.get_sg_id(
-                    sg['name']))
-            except (nt_exceptions.NeutronClientException,
-                    NotFound) as e:
-                print "Security group %s failed to delete: %s" % (sg['name'],
-                                                                  repr(e))
-
-    def clean_roles(self):
-        roles_names = [role['name'] for role in self.config.roles]
-        for role in self.keystoneclient.roles.list():
-            if role.name not in roles_names:
-                continue
-            self.keystoneclient.roles.delete(self.get_role_id(role.name))
-            print('Role "%s" has been deleted' % role.name)
-
-    def clean_keypairs(self):
-        def delete_user_keypairs(_user):
-            if not _user.get('enabled'):
-                return
-            try:
-                self.switch_user(user=_user['name'], tenant=_user['tenant'],
-                                 password=_user['password'])
-            except ks_exceptions.Unauthorized:
-                return
-
-            keypairs = [k.id for k in self.novaclient.keypairs.list()]
-            if keypairs:
-                for key_pair in keypairs:
-                    self.novaclient.keypairs.delete(key_pair)
-            self.switch_user(user=self.username, password=self.password,
-                             tenant=self.tenant)
-
-        for user in self.config.users:
-            delete_user_keypairs(user)
-
-    def clean_users(self):
-        users_names = [user['name'] for user in self.config.users]
-        for user in self.keystoneclient.users.list():
-            if user.name not in users_names:
-                continue
-            self.keystoneclient.users.delete(self.get_user_id(user.name))
-            print('User "%s" has been deleted' % user.name)
-
-    def clean_tenants(self):
-        tenants_names = [tenant['name'] for tenant in self.config.tenants]
-        for tenant in self.keystoneclient.tenants.list():
-            if tenant.name not in tenants_names:
-                continue
-            self.keystoneclient.tenants.delete(self.get_tenant_id(tenant.name))
-            print('Tenant "%s" has been deleted' % tenant.name)
-
-    def clean_cinder_snapshots(self):
-        snapshots = self.config.cinder_snapshots
-        snapshots += itertools.chain(
-            *[tenant['cinder_snapshots'] for tenant in self.config.tenants
-              if 'cinder_snapshots' in tenant])
-        sn_names = [snapshot['name'] for snapshot in snapshots]
-        snapshots = self.cinderclient.volume_snapshots.list(
-            search_opts={'all_tenants': 1})
-        for snapshot in snapshots:
-            if snapshot.name not in sn_names:
-                continue
-            self.cinderclient.volume_snapshots.delete(
-                self.get_volume_snapshot_id(snapshot.display_name))
-            print('Snapshot "%s" has been deleted' % snapshot.display_name)
-
-    def clean_namespaces(self):
-        ip_addr = self.get_vagrant_vm_ip()
-        if self.openstack_release == 'grizzly':
-            cmd = 'quantum-netns-cleanup'
-        else:
-            cmd = 'neutron-netns-cleanup'
-        self.migration_utils.execute_command_on_vm(ip_addr, cmd, 'root')
-
-    def clean_lb_pools(self):
-        pools = getattr(self.config, 'pools', [])
-        pools += itertools.chain(
-            *[tenant['pools'] for tenant in self.config.tenants
-              if 'pools' in tenant])
-        pools_names = [pool['name'] for pool in pools]
-        for pool in self.neutronclient.list_pools()['pools']:
-            if pool['name'] in pools_names:
-                self.neutronclient.delete_pool(pool['id'])
-                print('LBaaS pool "%s" has been deleted' % pool['name'])
-
-    def clean_lb_vips(self):
-        vips = getattr(self.config, 'vips', [])
-        vips += itertools.chain(
-            *[tenant['vips'] for tenant in self.config.tenants
-              if 'vips' in tenant])
-        vips_names = [vip['name'] for vip in vips]
-        for vip in self.neutronclient.list_vips()['vips']:
-            if vip['name'] in vips_names:
-                self.neutronclient.delete_vip(vip['id'])
-                print('LBaaS vip "%s" has been deleted' % vip['name'])
-
-    def clean_lb_members(self):
-        members = getattr(self.config, 'lb_members', [])
-        members += itertools.chain(
-            *[tenant['members'] for tenant in self.config.tenants
-              if 'members' in tenant])
-        member_address = [member['address'] for member in members]
-        for member in self.neutronclient.list_members()['members']:
-            if member['address'] in member_address:
-                self.neutronclient.delete_member(member['id'])
-                msg = 'LBaaS member for tenant "%s" has been deleted'
-                print(msg % member['tenant_id'])
-
-    def clean_lbaas_health_monitors(self):
-        def check_tenant(tenant_id):
-            try:
-                self.keystoneclient.tenants.get(tenant_id)
-                return True
-            except ks_exceptions.NotFound:
-                return False
-
-        tenants_ids = [self.get_tenant_id(tenant['name'])
-                       for tenant in self.config.tenants
-                       if self.tenant_exists(tenant['name'])]
-        tenants_ids.append(self.get_tenant_id(self.tenant))
-        for hm in self.neutronclient.list_health_monitors()['health_monitors']:
-            if hm['tenant_id'] in tenants_ids or not check_tenant(
-                    hm['tenant_id']):
-                self.neutronclient.delete_health_monitor(hm['id'])
-                msg = 'LBaaS health monitor for tenant "%s" has been deleted'
-                print(msg % hm['tenant_id'])
-
-    def clean_server_groups(self):
-        server_group_from_config = self.config.server_groups
-        for tenant in self.config.tenants:
-            if tenant.get('server_groups'):
-                server_group_from_config.extend(tenant['server_groups'])
-        server_group_names = [sg['name'] for sg in server_group_from_config]
-        for server_group in self.get_all_server_groups():
-            if server_group.name in server_group_names:
-                self.novaclient.server_groups.delete(server_group.id)
-
-    def clean_all_networking(self):
-        self.clean_fips()
-        self.clean_lb_members()
-        self.clean_lb_vips()
-        self.clean_lb_pools()
-        self.clean_lbaas_health_monitors()
-        self.clean_routers()
-        self.clean_networks()
-        self.clean_namespaces()
-
-    def clean_objects(self):
-        self.clean_vms()
-        self.clean_flavors()
-        self.clean_images()
-        self.clean_snapshots()
-        self.clean_cinder_snapshots()
-        self.clean_volumes()
-        self.clean_all_networking()
-        self.clean_security_groups()
-        self.clean_roles()
-        self.clean_keypairs()
-        if self.server_groups_supported:
-            self.clean_server_groups()
-        self.clean_users()
-        self.clean_tenants()
 
 
 if __name__ == '__main__':
