+++ devlab/tests/testcases/test_verify_dst_del_tenant_resources.py
